requirements.txt
Flask>=3.0,<4
python-dotenv>=1.0,<2
pandas>=2.0,<3
matplotlib>=3.8,<4
